refactor: extract state formatting helpers
