+++ CONTRIBUTING.md
@@ -4,6 +4,7 @@
 
 - `src/room-card.js` is the canonical source file.
 - `src/lib/values.js` owns browser-independent number/string helpers.
+- `src/lib/formatting.js` owns HA state/attribute formatting and existing fallbacks.
 - `dist/room-card.js` is the generated HACS artifact and must not be edited directly.
 - `dist/rooms/` contains the additional assets installed by HACS.
 - `tests/` contains automated regression tests.
